Split communicability_alg to subgraph.py and communicability.py
