+++ doc/source/reference/algorithms.centrality.rst
@@ -57,16 +57,14 @@ Eigenvector
    katz_centrality
    katz_centrality_numpy
 
-Communicability
+Subgraph
 ---------------
 .. autosummary::
    :toctree: generated/
 
-   communicability
-   communicability_exp
-   communicability_centrality
-   communicability_centrality_exp
-   communicability_betweenness_centrality
+   subgraph_centrality
+   subgraph_centrality_exp
+   subgraph_betweenness_centrality
    estrada_index
 
 Load
