+++ doc/source/reference/algorithms.communicability.rst
@@ -1,0 +1,11 @@
+***************
+communicability
+***************
+
+.. automodule:: networkx.algorithms.communicability
+.. autosummary::
+   :toctree: generated/
+
+   communicability
+   communicability_exp
+
